fix(mols): select per-order coprime multipliers with even-order fallback
- molsMultipliers picks m1, m2 per grid order so m1, m2, and m1-m2 are
  all coprime to the order; even orders (incl. 2 and 6) fall back to the
  single-square (1,1) score, keeping rows deterministic and duplicate-free
- correct hashToGridIndex comment: the hash is stable, the folded index
  is recomputed per current grid order
- widen telemetry ClientHash to uint32 so folded indexes from pools
  larger than 256 remain representable

+++ portal/telemetry/trace.go
@@ -11,7 +11,7 @@ type SelectionTrace struct {
 	// at selection time. Used only for sampled debug-log correlation. Not a
 	// Prometheus label (would unbounded cardinality) and not a public field on
 	// any external API.
-	ClientHash uint8
+	ClientHash uint32
 
 	// Mode is "priority" or "multihop", matching the calling method.
 	Mode string
